pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "splatsdb-ui"
version = "0.1.0"
description = "SplatsDB Vector Search Engine — Professional Desktop UI"
readme = "README.md"
license = {text = "GPL-3.0"}
requires-python = ">=3.10"
authors = [
    {name = "Brian Schwabauer", email = "schwabauerbriantomas@gmail.com"},
]

dependencies = [
    "PySide6>=6.6",
    "httpx>=0.27",
    "pillow>=10.0",
    "numpy>=1.26",
    "pyperclip>=1.8",
]

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3",
    "paddleocr>=2.7",
    "paddlepaddle>=2.5",
]
embeddings = [
    "torch>=2.1",
    "transformers>=4.36",
    "sentence-transformers>=2.2",
    "onnxruntime>=1.16",
]
all = [
    "splatsdb-ui[ocr,embeddings]",
]
dev = [
    "pytest>=7.0",
    "pytest-qt>=4.2",
    "ruff>=0.1",
]

[project.scripts]
splatsdb-ui = "splatsdb_ui.main:run"

[project.gui-scripts]
splatsdb-ui = "splatsdb_ui.main:run"

[tool.setuptools.packages.find]
include = ["splatsdb_ui*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
